Move util functions under utils.go
Given all such functions are already under utils.go,
it's natural to put the remaining two there as well.
